cli: Move file-scope resolution out of parser
CLI parsing now delegates resolved file action execution to command-layer
file-scope helpers. The argument parser still owns file-scope data types
and live and batch resolver functions.

The parser module mixes parser construction with repository-aware file
matching and batch metadata lookup. That keeps a shared command concern in
the largest CLI module after command execution moved out.

This commit addresses that by moving the file-scope data types and live and
batch scope resolvers into `cli.file_scope`, leaving the parser to import
the resolver interface.

The next commit will move resolver tests to the new module boundary.
