+++ src/git_stage_batch/cli/argument_parser.py
@@ -7,10 +7,7 @@ import os
 import shlex
 import sys
 import tempfile
-from collections.abc import Sequence
 from contextlib import nullcontext
-from dataclasses import dataclass
-from enum import Enum
 from importlib import resources
 from pathlib import Path
 
@@ -26,69 +23,16 @@ from ..commands.file_scope.multi_file_actions import (
     run_for_each_resolved_file,
     skip_each_resolved_file,
 )
-from ..data.file_tracking import list_untracked_files
 from ..exceptions import CommandError
 from ..i18n import _
 from ..utils.command import run_command
-from ..utils.file_patterns import (
-    list_changed_files,
-    list_staged_files,
-    resolve_gitignore_style_patterns,
-)
 from ..utils.git import run_git_command
 from .completion import command_complete_files
-
-FileArgument = str | list[str] | None
-
-
-class FileScopeKind(str, Enum):
-    """How a command's optional file scope was requested."""
-
-    IMPLICIT = "implicit"
-    EXPLICIT = "explicit"
-    PATTERN = "pattern"
-
-
-@dataclass(frozen=True)
-class FileScope:
-    """Resolved command file scope with explicit origin and concrete files."""
-
-    kind: FileScopeKind
-    files: tuple[str, ...] = ()
-
-    @classmethod
-    def implicit(cls) -> "FileScope":
-        return cls(FileScopeKind.IMPLICIT)
-
-    @classmethod
-    def explicit(cls, file_path: str) -> "FileScope":
-        return cls(FileScopeKind.EXPLICIT, (file_path,))
-
-    @classmethod
-    def pattern(cls, files: list[str]) -> "FileScope":
-        return cls(FileScopeKind.PATTERN, tuple(files))
-
-    @property
-    def is_implicit(self) -> bool:
-        return self.kind == FileScopeKind.IMPLICIT
-
-    @property
-    def is_multiple(self) -> bool:
-        return len(self.files) > 1
-
-    def optional_file(self) -> str | None:
-        """Return the single file path for this scope, or None for implicit scope."""
-        if self.is_implicit:
-            return None
-        if self.is_multiple:
-            raise ValueError("multiple file scope cannot be represented by one path")
-        return self.files[0]
-
-    def require_single_file(self, error_message: str) -> str | None:
-        """Return an optional single file path, or raise for a multi-file scope."""
-        if self.is_multiple:
-            raise CommandError(error_message)
-        return self.optional_file()
+from .file_scope import (
+    FileArgument,
+    resolve_batch_file_scope,
+    resolve_live_file_scope,
+)
 
 
 class GitHelpArgumentParser(argparse.ArgumentParser):
@@ -312,135 +256,6 @@ def _normalize_parsed_file_arguments(args: argparse.Namespace) -> None:
         args.file = file_values[0]
     else:
         args.file = file_values
-
-
-def _resolve_file_patterns(
-    file_arg: FileArgument,
-    file_patterns: list[str] | None,
-) -> list[str] | None:
-    """Return combined pattern arguments, preserving pathless --file."""
-    if file_arg == "":
-        if file_patterns is not None:
-            raise CommandError(_("Cannot use --file together with --files."))
-        return None
-
-    patterns: list[str] = []
-    if isinstance(file_arg, str):
-        patterns.append(file_arg)
-    elif file_arg is not None:
-        patterns.extend(file_arg)
-
-    if file_patterns is not None:
-        patterns.extend(file_patterns)
-
-    return patterns or None
-
-
-def _file_arg_values(file_arg: FileArgument) -> list[str]:
-    """Return argument-bearing --file values."""
-    if file_arg is None or file_arg == "":
-        return []
-    if isinstance(file_arg, str):
-        return [file_arg]
-    return list(file_arg)
-
-
-def _normalize_file_argument_path(path: str) -> str:
-    """Normalize a user-provided file path for exact candidate lookup."""
-    normalized = path.replace("\\", "/")
-    if normalized.startswith("./"):
-        normalized = normalized[2:]
-    return normalized
-
-
-def _resolve_file_argument_patterns(
-    candidates: Sequence[str],
-    file_arg: FileArgument,
-    file_patterns: list[str] | None,
-) -> tuple[list[str], list[str]]:
-    """Resolve --file/--files values against candidates with exact --file fallback."""
-    file_values = _file_arg_values(file_arg)
-    display_patterns = [*file_values, *(file_patterns or [])]
-    candidate_by_path = {
-        _normalize_file_argument_path(candidate): candidate
-        for candidate in candidates
-    }
-
-    exact_files: list[str] = []
-    pattern_values: list[str] = []
-    for value in file_values:
-        exact_file = candidate_by_path.get(_normalize_file_argument_path(value))
-        if exact_file is not None:
-            exact_files.append(exact_file)
-        else:
-            pattern_values.append(value)
-
-    pattern_values.extend(file_patterns or [])
-    resolved_patterns = (
-        resolve_gitignore_style_patterns(candidates, pattern_values)
-        if pattern_values else
-        []
-    )
-    resolved_files = list(dict.fromkeys([*exact_files, *resolved_patterns]))
-    return resolved_files, display_patterns
-
-
-def _resolve_live_file_scope(
-    file_arg: FileArgument,
-    file_patterns: list[str] | None,
-    *,
-    include_staged: bool = False,
-) -> FileScope:
-    """Resolve single-file or pattern-based live file scope."""
-    resolved_patterns = _resolve_file_patterns(file_arg, file_patterns)
-    if resolved_patterns is None:
-        return FileScope.implicit() if file_arg is None else FileScope.explicit("")
-
-    candidate_files = [*list_changed_files(), *list_untracked_files()]
-    if include_staged:
-        candidate_files.extend(list_staged_files())
-    candidate_files = list(dict.fromkeys(candidate_files))
-    resolved_files, display_patterns = _resolve_file_argument_patterns(
-        candidate_files,
-        file_arg,
-        file_patterns,
-    )
-    if not resolved_files:
-        raise CommandError(
-            _("No changed files matched: {patterns}").format(
-                patterns=", ".join(display_patterns),
-            )
-        )
-    return FileScope.pattern(resolved_files)
-
-
-def _resolve_batch_file_scope(
-    batch_name: str,
-    file_arg: FileArgument,
-    file_patterns: list[str] | None,
-) -> FileScope:
-    """Resolve single-file or pattern-based batch file scope."""
-    lookup_batch_name = batch_name_for_source_lookup(batch_name)
-    resolved_patterns = _resolve_file_patterns(file_arg, file_patterns)
-    if resolved_patterns is None:
-        return FileScope.implicit() if file_arg is None else FileScope.explicit("")
-    if not batch_exists(lookup_batch_name):
-        raise CommandError(_("Batch '{name}' does not exist").format(name=lookup_batch_name))
-
-    metadata = read_batch_metadata(lookup_batch_name)
-    resolved_files, display_patterns = _resolve_file_argument_patterns(
-        list(metadata.get("files", {}).keys()),
-        file_arg,
-        file_patterns,
-    )
-    if not resolved_files:
-        raise CommandError(
-            _("No files in batch '{name}' matched: {patterns}").format(
-                name=lookup_batch_name,
-                patterns=", ".join(display_patterns),
-            )
-        )
-    return FileScope.pattern(resolved_files)
 
 
 def _resolve_replacement_text(args: argparse.Namespace) -> str | None:
@@ -675,9 +490,9 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
         if args.as_text is not None and args.as_stdin:
             raise CommandError(_("Cannot use `--as` and `--as-stdin` together."))
         resolved_file_scope = (
-            _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+            resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
             if args.from_batch
-            else _resolve_live_file_scope(args.file, args.file_patterns)
+            else resolve_live_file_scope(args.file, args.file_patterns)
         )
         if args.page is not None:
             lookup_batch = batch_name_for_source_lookup(args.from_batch) if args.from_batch else None
@@ -857,7 +672,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
             if args.line_ids and args.from_batch and not args.to_batch:
                 if args.no_edge_overlap:
                     raise CommandError(_("`--no-edge-overlap` only applies to live `include --line --as` operations."))
-                resolved_batch_scope = _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+                resolved_batch_scope = resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
                 resolved_file = resolved_batch_scope.require_single_file(_("Cannot use --lines with multiple files."))
                 replacement_text = _resolve_replacement_text(args)
                 commands.command_include_from_batch(
@@ -868,7 +683,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 )
                 return
             if args.line_ids and not args.from_batch and not args.to_batch:
-                resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+                resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
                 resolved_file = resolved_live_scope.require_single_file(_("Cannot use --lines with multiple files."))
                 replacement_text = _resolve_replacement_text(args)
                 commands.command_include_line_as(
@@ -884,7 +699,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 and args.from_batch is None
                 and args.to_batch is None
             ):
-                resolved_live_scope = _resolve_live_file_scope(
+                resolved_live_scope = resolve_live_file_scope(
                     args.file,
                     args.file_patterns,
                     include_staged=True,
@@ -910,7 +725,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
         if args.no_edge_overlap:
             raise CommandError(_("`--no-edge-overlap` requires `include --line --as`."))
         if args.from_batch:
-            resolved_batch_scope = _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+            resolved_batch_scope = resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
             run_for_each_resolved_file(
                 resolved_batch_scope,
                 lambda file: commands.command_include_from_batch(args.from_batch, args.line_ids, file),
@@ -919,7 +734,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 worktree_paths=resolved_batch_scope.files,
             )
         elif args.to_batch:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             run_for_each_resolved_file(
                 resolved_live_scope,
                 lambda file: commands.command_include_to_batch(
@@ -932,7 +747,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 undo_operation=f"include --to {shlex.quote(args.to_batch)}",
             )
         elif args.line_ids:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             resolved_file = resolved_live_scope.require_single_file(_("Cannot use --lines with multiple files."))
             commands.command_include_line(
                 args.line_ids,
@@ -940,7 +755,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 auto_advance=args.auto_advance,
             )
         else:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             if resolved_live_scope.is_multiple:
                 include_each_resolved_file(
                     list(resolved_live_scope.files),
@@ -979,7 +794,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
     _add_auto_advance_arguments(parser_skip)
 
     def dispatch_skip(args: argparse.Namespace) -> None:
-        resolved_file_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+        resolved_file_scope = resolve_live_file_scope(args.file, args.file_patterns)
         if args.line_ids:
             resolved_file = resolved_file_scope.require_single_file(_("Cannot use --lines with multiple files."))
             commands.command_skip_line(
@@ -1068,7 +883,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
             if args.as_text is not None and args.as_stdin:
                 raise CommandError(_("Cannot use `--as` and `--as-stdin` together."))
             if args.to_batch and args.line_ids and not args.from_batch:
-                resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+                resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
                 resolved_file = resolved_live_scope.require_single_file(_("Cannot use --lines with multiple files."))
                 replacement_text = _resolve_replacement_text(args)
                 commands.command_discard_line_as_to_batch(
@@ -1085,7 +900,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 and args.from_batch is None
                 and args.line_ids is None
             ):
-                resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+                resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
                 if resolved_live_scope.is_implicit:
                     raise CommandError(
                         _("`discard --as` requires `--file`, or `--to` with `--line`.")
@@ -1107,7 +922,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
         if args.no_edge_overlap:
             raise CommandError(_("`--no-edge-overlap` requires `discard --to --line --as`."))
         if args.from_batch:
-            resolved_batch_scope = _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+            resolved_batch_scope = resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
             run_for_each_resolved_file(
                 resolved_batch_scope,
                 lambda file: commands.command_discard_from_batch(args.from_batch, args.line_ids, file),
@@ -1116,7 +931,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 worktree_paths=resolved_batch_scope.files,
             )
         elif args.to_batch:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             if resolved_live_scope.is_multiple and args.line_ids is None:
                 discard_to_batch_each_resolved_file(
                     args.to_batch,
@@ -1137,7 +952,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                     worktree_paths=resolved_live_scope.files,
                 )
         elif args.line_ids:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             resolved_file = resolved_live_scope.require_single_file(_("Cannot use --lines with multiple files."))
             commands.command_discard_line(
                 args.line_ids,
@@ -1145,7 +960,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
                 auto_advance=args.auto_advance,
             )
         else:
-            resolved_live_scope = _resolve_live_file_scope(args.file, args.file_patterns)
+            resolved_live_scope = resolve_live_file_scope(args.file, args.file_patterns)
             if not resolved_live_scope.is_implicit:
                 run_for_each_resolved_file(
                     resolved_live_scope,
@@ -1335,7 +1150,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
     )
 
     def dispatch_apply(args: argparse.Namespace) -> None:
-        resolved_file_scope = _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+        resolved_file_scope = resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
         run_for_each_resolved_file(
             resolved_file_scope,
             lambda file: commands.command_apply_from_batch(
@@ -1384,7 +1199,7 @@ def parse_command_line(args: list[str], *, quiet: bool = False) -> argparse.Name
     )
 
     def dispatch_reset(args: argparse.Namespace) -> None:
-        resolved_file_scope = _resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
+        resolved_file_scope = resolve_batch_file_scope(args.from_batch, args.file, args.file_patterns)
         command_parts = ["reset", "--from", shlex.quote(args.from_batch)]
         if args.to_batch is not None:
             command_parts.extend(["--to", shlex.quote(args.to_batch)])
